geo/wkt: divide WKT parser into multiple files
Release note: None
